refactor: explicit imports instead of wildcard, remove editor-related files
