+++ .gitignore
@@ -199,3 +199,6 @@ node_modules
 
 # https://github.com/octokit/webhooks
 webhooks
+
+# VS Code-related stuff
+.vscode
