--- src/.vscode/extensions.json
@@ -1,3 +1,0 @@
-{
-  "recommendations": ["charliermarsh.ruff"]
-}
