--- src/.vscode/settings.json
@@ -1,10 +1,0 @@
-{
-  "editor.codeActionsOnSave": {
-    "source.unusedImports": "always"
-  },
-  "[python]": {
-    "editor.tabSize": 2,
-    "editor.indentSize": "tabSize",
-    "editor.defaultFormatter": "charliermarsh.ruff"
-  }
-}
