+++ src/backend/app/dependencies/__init__.py
@@ -1,1 +1,1 @@
-from .rate_limit import *  # noqa
+from .rate_limit import RateLimitDep  # noqa: F401
