+++ src/backend/app/github/webhook/schemas/__init__.py
@@ -1,2 +1,2 @@
-from .events import *  # noqa
-from .headers import *  # noqa
+from .events import PushEvent, ReleaseEvent, RepositoryEvent, WebhookEvent  # noqa: F401
+from .headers import WebhookHeaders  # noqa: F401
